requirements.txt
# FastAPI Framework
fastapi==0.131.0
uvicorn[standard]==0.41.0

# Database
SQLAlchemy==2.0.46
psycopg2-binary==2.9.11

# Data Validation
pydantic==2.12.5
pydantic-settings==2.13.1
email-validator==2.3.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

